api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile api/requirements.in
#
aniso8601==10.0.1
    # via flask-restful
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via flask
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
click==8.3.1
    # via flask
distro==1.9.0
    # via openai
flake8==7.3.0
    # via -r api/requirements.in
flask==3.1.3
    # via
    #   -r api/requirements.in
    #   flask-cors
    #   flask-jwt-extended
    #   flask-restful
    #   flask-sqlalchemy
flask-cors==6.0.2
    # via -r api/requirements.in
flask-jwt-extended==4.7.1
    # via -r api/requirements.in
flask-restful==0.3.10
    # via -r api/requirements.in
flask-sqlalchemy==3.1.1
    # via -r api/requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
imutils==0.5.4
    # via -r api/requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jiter==0.13.0
    # via openai
jsonschema==4.26.0
    # via -r api/requirements.in
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via flake8
mysql-connector-python==9.6.0
    # via -r api/requirements.in
numpy==2.4.2
    # via -r api/requirements.in
openai==2.24.0
    # via -r api/requirements.in
pycodestyle==2.14.0
    # via flake8
pydantic==2.12.5
    # via openai
pydantic-core==2.41.5
    # via pydantic
pyflakes==3.4.0
    # via flake8
pyjwt==2.11.0
    # via flask-jwt-extended
python-dotenv==1.2.2
    # via -r api/requirements.in
pytz==2025.2
    # via flask-restful
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via flask-restful
sniffio==1.3.1
    # via openai
sqlalchemy==2.0.48
    # via flask-sqlalchemy
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uwsgi==2.0.31
    # via -r api/requirements.in
views==0.3
    # via -r api/requirements.in
werkzeug==3.1.6
    # via
    #   flask
    #   flask-cors
    #   flask-jwt-extended
